CharacterRecoginition/test-project/requirements.txt
tensorflow
numpy
matplotlib
Pillow
Flask
